add validation

+++ src/main/java/com/springboot/fullstack/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.springboot.fullstack.student;
 
+import com.springboot.fullstack.student.exception.BadRequestException;
+import com.springboot.fullstack.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,12 +18,22 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        // check if email is taken
+        Boolean emailExists = studentRepository.selectEmailExists(student.getEmail());
+
+        if (emailExists) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " taken"
+            );
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
-        // check if student exists
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exist."
+            );
+        }
         studentRepository.deleteById(studentId);
     }
 }
